Initial implementation
This simple tool will take almost any virtual disk and quickly generate
a container for it.  It can generate a "full" image, or if you specify
an existing base docker image, it will attempt to create a new image
layer containing only the diffrences between the virtual disk and the
docker image.